fix: (platform) Bug fix for multiple menu triggers with same menu
For a scenario where a single menu is referenced by multiple trigger via
the `menuTriggerFor` directive, there is a bug when clicking from one
button to the next. Clicking on the first button will open the menu.
Clicking on the second button will then close the menu, without opening
it again and attaching to second trigger element. This happens because
the close menu event from an "outside" click is resolved after the open
menu call from the second button click, so the menu does not open the
second time.

I added a `setTimeout` delay on the open menu call to insure it happens
after the close menu is resolved.

Fixing unit tests after adding delay to openMenu

CSS style fixes